Split up into PSR-4 format and fix tests. (Including kill off 'lib' folder, move phpunit.xml and update composer)
