+++ composer.json
@@ -31,7 +31,7 @@
   "require": {
     "php": ">=5.3.0",
     "ext-mbstring": "*",
-    "ext-http": "*"
+    "ext-http": "~2.0"
   },
   "require-dev": {
     "phpunit/phpunit": "~4.8",
@@ -39,7 +39,7 @@
   },
   "autoload": {
     "psr-4": {
-      "ZipStreamer": "src/"
+      "Rivimey\\ZipStreamer\\": "src/"
     }
   }
 }
