+++ src/Count64/Count64.php
@@ -1,0 +1,33 @@
+<?php
+/**
+ * Simple class to support some very basic operations on 64 bit intergers
+ * on 32 bit machines.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * @author Nicolai Ehemann <en@enlightened.de>
+ * @copyright Copyright (C) 2013-2014 Nicolai Ehemann and contributors
+ * @license GNU GPL
+ */
+namespace Rivimey\ZipStreamer\Count64;
+
+abstract class Count64 {
+  public static function construct($value = 0, $limit32Bit = False) {
+    if (4 == PHP_INT_SIZE) {
+      return new Count64_32($value, $limit32Bit);
+    } else {
+      return new Count64_64($value, $limit32Bit);
+    }
+  }
+}
