--- test/phpunit.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<phpunit
-     colors="true"
-     strict="true"
-     verbose="true"
-     debug="true"
->
-  <testsuite name="ZipStreamer">
-    <directory>./</directory>
-  </testsuite>
-</phpunit>
